+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Fixed
 * Fix issue in kwcoco subset where paths were invalid if written outside of
   the source kwcoco bundle.
+* Remove the gdal cache, which caused issues in read/write cases.
 
 
 ## Version 0.2.30 - Released 2022-04-19